+++ connection.go
@@ -47,10 +47,7 @@ import (
 //
 // This interface is central part for access to various systems
 // embedded to ydb through one configured connection method.
-type Connection = *connection
-
-//nolint:maligned
-type connection struct {
+type Connection struct { //nolint:maligned
 	userInfo *dsn.UserInfo
 
 	opts []Option
@@ -93,15 +90,15 @@ type connection struct {
 	mtx      sync.Mutex
 	balancer *balancer.Balancer
 
-	children    map[uint64]Connection
+	children    map[uint64]*Connection
 	childrenMtx xsync.Mutex
-	onClose     []func(c *connection)
+	onClose     []func(c *Connection)
 
 	panicCallback func(e interface{})
 }
 
 // Close closes connection and clear resources
-func (c *connection) Close(ctx context.Context) error {
+func (c *Connection) Close(ctx context.Context) error {
 	c.mtx.Lock()
 	defer c.mtx.Unlock()
 
@@ -145,7 +142,7 @@ func (c *connection) Close(ctx context.Context) error {
 	return nil
 }
 
-func (c *connection) Invoke(
+func (c *Connection) Invoke(
 	ctx context.Context,
 	method string,
 	args interface{},
@@ -161,7 +158,7 @@ func (c *connection) Invoke(
 	)
 }
 
-func (c *connection) NewStream(
+func (c *Connection) NewStream(
 	ctx context.Context,
 	desc *grpc.StreamDesc,
 	method string,
@@ -176,22 +173,22 @@ func (c *connection) NewStream(
 }
 
 // Endpoint returns initial endpoint
-func (c *connection) Endpoint() string {
+func (c *Connection) Endpoint() string {
 	return c.config.Endpoint()
 }
 
 // Name returns database name
-func (c *connection) Name() string {
+func (c *Connection) Name() string {
 	return c.config.Database()
 }
 
 // Secure returns true if database connection is secure
-func (c *connection) Secure() bool {
+func (c *Connection) Secure() bool {
 	return c.config.Secure()
 }
 
 // Table returns table client
-func (c *connection) Table() table.Client {
+func (c *Connection) Table() table.Client {
 	c.tableOnce.Init(func() closeFunc {
 		c.table = internalTable.New(
 			c.balancer,
@@ -212,7 +209,7 @@ func (c *connection) Table() table.Client {
 }
 
 // Scheme returns scheme client
-func (c *connection) Scheme() scheme.Client {
+func (c *Connection) Scheme() scheme.Client {
 	c.schemeOnce.Init(func() closeFunc {
 		c.scheme = internalScheme.New(
 			c.balancer,
@@ -234,7 +231,7 @@ func (c *connection) Scheme() scheme.Client {
 }
 
 // Coordination returns coordination client
-func (c *connection) Coordination() coordination.Client {
+func (c *Connection) Coordination() coordination.Client {
 	c.coordinationOnce.Init(func() closeFunc {
 		c.coordination = internalCoordination.New(
 			c.balancer,
@@ -255,7 +252,7 @@ func (c *connection) Coordination() coordination.Client {
 }
 
 // Ratelimiter returns ratelimiter client
-func (c *connection) Ratelimiter() ratelimiter.Client {
+func (c *Connection) Ratelimiter() ratelimiter.Client {
 	c.ratelimiterOnce.Init(func() closeFunc {
 		c.ratelimiter = internalRatelimiter.New(
 			c.balancer,
@@ -276,7 +273,7 @@ func (c *connection) Ratelimiter() ratelimiter.Client {
 }
 
 // Discovery returns discovery client
-func (c *connection) Discovery() discovery.Client {
+func (c *Connection) Discovery() discovery.Client {
 	c.discoveryOnce.Init(func() closeFunc {
 		c.discovery = internalDiscovery.New(
 			c.pool.Get(endpoint.New(c.config.Endpoint())),
@@ -301,7 +298,7 @@ func (c *connection) Discovery() discovery.Client {
 }
 
 // Scripting returns scripting client
-func (c *connection) Scripting() scripting.Client {
+func (c *Connection) Scripting() scripting.Client {
 	c.scriptingOnce.Init(func() closeFunc {
 		c.scripting = internalScripting.New(
 			c.balancer,
@@ -322,7 +319,7 @@ func (c *connection) Scripting() scripting.Client {
 }
 
 // Topic returns topic client
-func (c *connection) Topic() topic.Client {
+func (c *Connection) Topic() topic.Client {
 	c.topicOnce.Init(func() closeFunc {
 		c.topic = topicclientinternal.New(c.balancer, c.config.Credentials(), c.topicOptions...)
 		return c.topic.Close
@@ -337,7 +334,7 @@ func (c *connection) Topic() topic.Client {
 //	"grpc[s]://{endpoint}/{database}[?param=value]"
 //
 // See sugar.DSN helper for make dsn from endpoint and database
-func Open(ctx context.Context, dsn string, opts ...Option) (_ Connection, err error) {
+func Open(ctx context.Context, dsn string, opts ...Option) (_ *Connection, err error) {
 	return open(
 		ctx,
 		append(
@@ -352,14 +349,14 @@ func Open(ctx context.Context, dsn string, opts ...Option) (_ Connection, err er
 // New connects to database and return driver runtime holder
 //
 // Deprecated: use Open with required param connectionString instead
-func New(ctx context.Context, opts ...Option) (_ Connection, err error) {
+func New(ctx context.Context, opts ...Option) (_ *Connection, err error) {
 	return open(ctx, opts...)
 }
 
-func newConnectionFromOptions(ctx context.Context, opts ...Option) (_ *connection, err error) {
-	c := &connection{
+func newConnectionFromOptions(ctx context.Context, opts ...Option) (_ *Connection, err error) {
+	c := &Connection{
 		opts:     opts,
-		children: make(map[uint64]Connection),
+		children: make(map[uint64]*Connection),
 	}
 	if caFile, has := os.LookupEnv("YDB_SSL_ROOT_CERTIFICATES_FILE"); has {
 		opts = append([]Option{WithCertificatesFromFile(caFile)}, opts...)
@@ -392,7 +389,7 @@ func newConnectionFromOptions(ctx context.Context, opts ...Option) (_ *connectio
 	return c, nil
 }
 
-func connect(ctx context.Context, c *connection) error {
+func connect(ctx context.Context, c *Connection) error {
 	var err error
 
 	if c.config.Endpoint() == "" {
@@ -435,7 +432,7 @@ func connect(ctx context.Context, c *connection) error {
 	return nil
 }
 
-func open(ctx context.Context, opts ...Option) (_ Connection, err error) {
+func open(ctx context.Context, opts ...Option) (_ *Connection, err error) {
 	c, err := newConnectionFromOptions(ctx, opts...)
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
@@ -451,7 +448,7 @@ func open(ctx context.Context, opts ...Option) (_ Connection, err error) {
 // unary and streaming RPC over internal driver balancer.
 //
 // Warning: for connect to driver-unsupported YDB services
-func GRPCConn(cc Connection) grpc.ClientConnInterface {
+func GRPCConn(cc *Connection) grpc.ClientConnInterface {
 	return cc
 }